Optimizations
Dictionary access removed and replaced with List Iteration in CsvMapping
class.

Use a compiled lambda for creating a new instance of an object in
CsvMapping class

Returning a CsvPropertyMapping in the MapProperty Code, so a custom
converter can be set, so we can override the default one.

Added Performance Tests to benchmark the project