Solve Issue 5: Tokenizer
This commit introduces the concept of a Tokenizer. It enables a user to
inject a custom tokenizer for processing a CSV line. This allows maximum
flexibility over the format.

Two sample Tokenizers are added. One is simple StringSplitTokenizer,
which simply splits a String. The other one is a more complex Tokenizer
based on Regular Expression.

This commit also resolves Issue #4 in a clean way, so that no
preprocessing is required to read a CSV file.