Initial Documentation
I have decided to switch the GitHub Wiki with a custom built
documentation based on the Sphinx project. Sphinx has a great theme:
Read The Docs and it makes publishing a documentation to various formats
really easy.

It is very easy to configure and more enjoyable, than a GitHub Wiki. I
am not sure yet, where to publish the final documentation (GitHub Pages,
Self Hosted?).